+++ MicroERP.Data/MicroERP.Data.Api/Repositories/ApiInvoiceRepository.cs
@@ -101,7 +101,7 @@ namespace MicroERP.Data.Api.Repositories
             }
         }
 
-        public async Task<IEnumerable<InvoiceModel>> Search(int? customerID = null, DateTime? begin = null, DateTime? end = null, double? minTotal = null, double? maxTotal = null)
+        public async Task<IEnumerable<InvoiceModel>> Search(int? customerID = null, DateTime? begin = null, DateTime? end = null, decimal? minTotal = null, decimal? maxTotal = null)
         {
             var url = new StringBuilder(this.ConnectionString).Append("/invoices?");
